creation fichier pourr test relais

+++ scripts/refroidissement/on.py
@@ -6,6 +6,7 @@ from time import sleep
 sys.path.insert(1, '/home/pi/Desktop/www/aqua/scripts/')
 import functions
 import RPi.GPIO as GPIO
+import os
 
 relais = 20
 
@@ -13,6 +14,7 @@ GPIO.setmode(GPIO.BCM)
 GPIO.setwarnings(False)
 functions.onled(relais)
 
-f = open("/home/pi/Desktop/www/aqua/statusRelais/refroidissement", "w")
+if os.path.exists("/home/pi/Desktop/www/aqua/statusRelais/refroidissement") is False:
+    f = open("/home/pi/Desktop/www/aqua/statusRelais/refroidissement", "w")
 
 sys.exit()
